+++ src/auth/auth.module.ts
@@ -7,8 +7,7 @@ import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtStrategy } from './jwt.strategy';
 import { UsersModule } from 'src/users/users.module';
 import { AgentProfileModule } from 'src/agent_profile/agent_profile.module';
-import { SuperAgentProfileModule } from 'src/super_agent_profile/super_agent_profile.module';
-import { InvitationsModule } from 'src/invitations/invitations.module';
+import { InvitationsModule } from 'src/referals/referal.module';
 import { OtpService } from 'src/mailing/otp.mail';
 import { EmailService } from 'src/configs/email.helper';
 import { ResetPasswordService } from '../mailing/resetPassword.mail';
@@ -34,7 +33,6 @@ import { WalletModule } from 'src/wallet/wallet.module';
       }),
     }),
     InvitationsModule,
-    SuperAgentProfileModule,
     AgentProfileModule,
     UsersModule,
     PassportModule,
