+++ src/auth/auth.service.ts
@@ -6,14 +6,12 @@ import {
 import { JwtService } from '@nestjs/jwt';
 import { UserRepository } from 'src/entity/repositories/user.repo';
 import { AgentProfileRepository } from 'src/entity/repositories/agent_profile.repo';
-import { SuperAgentProfileRepository } from 'src/entity/repositories/super_agent_profile.repo';
 import { OtpService } from '../mailing/otp.mail';
 import { JwtPayload } from './jwt-payload';
 import * as bcrypt from 'bcryptjs';
 import { LoginUser } from 'src/users/users.entity';
 import { CreateAgentProfileDto } from 'src/agent_profile/agent_profile.dto';
 import { CreateUserDto } from 'src/users/users.dto';
-import { CreateSuperAgentProfileDto } from 'src/super_agent_profile/super_agent_profile.dto';
 import { UsersService } from 'src/users/users.service';
 import { ResetPasswordService } from '../mailing/resetPassword.mail';
 import { WalletService } from 'src/wallet/wallet.service';
@@ -24,7 +22,6 @@ export class AuthService {
     private readonly userRepo: UserRepository,
     private readonly userService: UsersService,
     private readonly agentRepo: AgentProfileRepository,
-    private readonly superAgentRepo: SuperAgentProfileRepository,
     private readonly walletService: WalletService,
     private readonly jwtService: JwtService,
     private readonly otpService: OtpService,
@@ -34,8 +31,7 @@ export class AuthService {
   async createUser(
     createUserDto:
       | CreateUserDto
-      | CreateAgentProfileDto
-      | CreateSuperAgentProfileDto,
+      | CreateAgentProfileDto,
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     req: any,
   ) {
@@ -45,17 +41,15 @@ export class AuthService {
       throw new BadRequestException('Email already exists');
     }
 
+    createUserDto.refCode = `${this.generateReferalCode()}-${createUserDto.firstname}`
+
     let user: any;
     if (role === 'Agent') {
       user = await this.createAgent(createUserDto as CreateAgentProfileDto);
-    } else if (role === 'SuperAgent') {
-      user = await this.createSuperAgent(
-        createUserDto as CreateSuperAgentProfileDto,
-      );
     } else {
       user = await this.userRepo.create(createUserDto);
     }
-
+    
     const otp = this.otpService.generateOTP();
     await this.otpService.sendOtpEmail(
       email,
@@ -78,17 +72,6 @@ export class AuthService {
     return { agentUser, agentProfile };
   }
 
-  private async createSuperAgent(
-    createSuperAgentDto: CreateSuperAgentProfileDto,
-  ) {
-    const superAgentUser = await this.userRepo.create(createSuperAgentDto);
-    const superAgentProfile = await this.superAgentRepo.create({
-      ...createSuperAgentDto,
-      user: superAgentUser._id,
-    });
-
-    return { superAgentUser, superAgentProfile };
-  }
 
   async validateUser(email: string, password: string) {
     const user = await this.userRepo.findOne({ email });
@@ -234,5 +217,19 @@ export class AuthService {
     newPassword: string,
   ) {
     return this.userService.updatePassword(userId, oldPassword, newPassword);
+  }
+
+
+  private generateReferalCode() {
+    const length = 10;
+    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    let result = '';
+  
+    for (let i = 0; i < length; i++) {
+      const generatedCode = Math.floor(Math.random() * characters.length);
+      result += characters[generatedCode];
+    }
+  
+    return result;
   }
 }
