+++ src/entity/index.ts
@@ -1,7 +1,6 @@
 export * from './entity.repositories';
 export * from './schema/user.schema';
 export * from './schema/agent_profile.schema';
-export * from './schema/super_agent_profile.schema';
 export * from './schema/invitation.schema';
 export * from './schema/wallet.schema';
 export * from './schema/category.schema';
