--- src/super_agent_profile/super_agent_profile.entity.ts
@@ -1,5 +1,0 @@
-export class SuperAgentProfile {
-  business_username: string;
-  region: string;
-  location: string;
-}
